ux(send-payment): auto focus input on popover opens
